mp_integer: move some tests from one file to the other.
